# doc/api.yaml
openapi: 3.1.0
info:
  version: 1.0.0
  title: WASA Photo REST API
  description: |
    This is the API specification for the WASA Photo REST API.

    It's part of the WASA Photo project for the Web and Software Architecture course at Sapienza University of Rome.

    The project description and details are availabe at [this link](http://gamificationlab.uniroma1.it/notes/Project.pdf)
servers:
  - url: http://localhost:8080
    description: Local development server
...
components:
  securitySchemes:
    bearerAuth:
      description: |
        This security scheme is used to authenticate a user.
        The user must be signed in to perform the request.
        The user must pass the bearer token in the Authorization header.
      type: http
      scheme: bearer
  schemas:
    username:
      description: Username of a user
      type: string
      minLength: 4
      maxLength: 20
      pattern: '^[a-zA-Z0-9_-]+$'
      readOnly: true
      example: mariorossi99
    resourceId:
      description: The Universally Unique Identifier (follows the standard)
      type: string
      format: uuid
      example: 550e8400-e29b-41d4-a716-446655440000
      readOnly: true
    date:
      description: representation of Date and time
      type: string
      format: date-time
      example: 2023-01-01T00:00:00Z
      readOnly: true
    caption:
      description: A caption for a post or a comment
      type: string
      minLength: 1
      maxLength: 800
      pattern: '.*'
      example: "This is a caption"
      readOnly: false
...
    counter:
      description: A simple counter,
                    keeps track of the number of followers, likes, etc.
      type: integer
      example: 8
      readOnly: true
    userCollection:
      description: A list of usersnames, 
                   it is used to keep track of the users that liked a post or a comment
      type: array
      minLength: 0
      maxLength: 69420
      items:
        anyOf:
          - $ref: '#/components/schemas/User/properties/username'
          - $ref: '#/components/schemas/User/properties/userId'
          - $ref: '#/components/schemas/User/properties/profileImage'
      readOnly: false
...
    Comment:
      title: Comment
      type: object
      description: A comment made by a user to a post
      properties:
        commentid:
          $ref: '#/components/schemas/resourceId'
        authorUsername:
          $ref: '#/components/schemas/username'
        authorId:
          $ref: '#/components/schemas/User/properties/userId'
        creationDate:
          $ref: '#/components/schemas/date'
        caption:
          $ref: '#/components/schemas/caption'
        likeCount:
          $ref: '#/components/schemas/counter'
      required:
        - commentId
        - authorId
        - authorUsername
        - creationDate
        - caption
...
    commentStream:
      description: An array of comments of a post
      type: array
      minLength: 0
      maxLength: 69420
      items:
        $ref: '#/components/schemas/Comment'
    Error:
      description: An error message
      type: object
      properties:
        message:
          type: string
          minLength: 1
          maxLength: 100
          pattern: '^[a-zA-Z0-9_-]$'
      required:
        - message
    Success:
      description: A successful request
      type: object
      properties:
        message:
          type: string
          minLength: 1
          maxLength: 100
          pattern: '^[a-zA-Z0-9_-]$'
        body:
          type: object
      required:
        - message
  parameters:
    userId:
      name: userId
      in: path
      description: The userId that is being requested
      required: true
      schema:
        $ref: '#/components/schemas/resourceId'
    postId:
      name: postId
      in: path
      description: The postId that is being requested
      required: true
      schema:
        $ref: '#/components/schemas/resourceId'
    commentId:
      name: commentId
      in: path
      description: The commentId that is being requested
      required: true
      schema:
        $ref: '#/components/schemas/resourceId'
    followingId:
      name: followingId
      in: path
      description: The userId that is being requested
      required: true
      schema:
        $ref: '#/components/schemas/resourceId'
    bannedId:
      name: bannedId
      in: path
      description: The userId that is being requested
      required: true
      schema:
        $ref: '#/components/schemas/resourceId'
...
  responses:
...
    NotFound: #for 404
      description: The requested resource was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized: #for 401
      description: The request requires user authentication, 
                    or the user is not authorized to perform the requested action
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    BadRequest: #for 400
      description: The request was not valid, 
                    the request body is missing or malformed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalServerError: #for 500
      description: An error occurred on the server side
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    Created: #for 201
      description: The resource was successfully created
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/resourceId'
    Ok: #for 200
      description: The request has succeeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Success'
...
    Comment: # 200 response for a comment
      description: A comment made by a user to a post
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Comment'
...
  requestBodies:
...
    Comment:
      description: A comment made by a user to a post
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Comment'
...
paths:
...
  /users/{userId}/posts/{postId}/comments:
    description: This endpoint handles the collection of comments of a post.
    parameters:
      - $ref: '#/components/parameters/userId'
      - $ref: '#/components/parameters/postId'
  
    post:
      tags: ["comment"]
      operationId: commentPhoto
      summary: Create a new comment
      description: |
        This request is used to create a new comment.
        The postownerId and the postId are passed as path parameters.
        The comment author is in the body.
        The comment details are passed in the request body.
        The response will retun the id of the new comment.
      requestBody:
        $ref: '#/components/requestBodies/Comment'
      responses:
        "201":
          $ref: '#/components/responses/Created'
        "400": #the request body is missing or malformed
          $ref: '#/components/responses/BadRequest'
        "404": #user not found
          $ref: '#/components/responses/NotFound'
        "500": #server error
          $ref: '#/components/responses/InternalServerError'
    
    get:
      tags: ["comment"]
      operationId: getPostComments
      summary: Get all the comments of a post
      description: |
        This endpoint is used to get all the comments of a post.
        The userId and the postId are passed
        The response will retun the list of comments of the post.
      responses:
        "200":
          description: The list of comments of the post
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/commentStream'
        "404": #post not found
          $ref: '#/components/responses/NotFound'
        "500": #server error
          $ref: '#/components/responses/InternalServerError'
  /users/{userId}/posts/{postId}/comments/{commentId}:
    description: This hendpoints handles comment edits and deletion
    parameters:
      - $ref: '#/components/parameters/userId'
      - $ref: '#/components/parameters/postId'
      - $ref: '#/components/parameters/commentId'
    
    get:
      tags: ["comment"]
      operationId: getComment
      summary: Get the details of a comment
      description: |
        This request is used to get the details of a comment.
        The postownerId, the postId and the commentId are passed as path parameters.
        The response will retun the details of the comment in the body.
      responses:
        "200":
          $ref: '#/components/responses/Comment'
        "404": #post not found
          $ref: '#/components/responses/NotFound'
        "401":
          $ref: '#/components/responses/Unauthorized'

    put: 
      tags: ["comment"]
      operationId: editComment
      summary: Edit a comment
      description: |
        This request is used to edit a comment.
        The postownerId, the postId and the commentId are passed as path parameters.
        A user can edit only is own comments. (owner is compared with the userId in the bearer token)
        The new comment details are passed in the request body.
        The response will retun the id of the new comment.
      requestBody:
        $ref: '#/components/requestBodies/Comment'
      responses:
        "202": #update successful
          $ref: '#/components/responses/Ok'
        "404":
          $ref: '#/components/responses/NotFound'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "500":
          $ref: '#/components/responses/InternalServerError'
  
    delete: 
      tags: ["comment"]
      operationId: uncommentPhoto
      summary: Delete a comment
      description:
        This request is used to delete a comment.
        The postownerId, the postId and the commentId are passed as path parameters.
        A user can delete only is own comments. (owner is compared with the userId in the bearer token)
        The response will retun the new collection of comments of the post.
      responses:
        "200":
          description: The new collection of comments of the post
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/commentStream'
        "404":
          $ref: '#/components/responses/NotFound'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "500":
          $ref: '#/components/responses/InternalServerError'
...
  /users/{userId}/following/{followingId}:
    description: A user followed by a user
    parameters:
      - $ref: '#/components/parameters/userId'
      - $ref: '#/components/parameters/followingId'

    put:
      tags: ["follow"]
      operationId: followUser
      summary: Follow a user
      description: |
        This request is used to follow a user.
        The userId of the user to follow is passed as a path parameter.
        The userId of the user who is following is taken from the bearer token.
        The response will retun the new list of followers of the user.
      responses:
        "201":
          description: The new list of followers of the user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/userCollection'
        "404": #user not found
          $ref: '#/components/responses/NotFound'
        "500": #server error
          $ref: '#/components/responses/InternalServerError'
        "401": #unauthorized
          $ref: '#/components/responses/Unauthorized'

    delete:
      tags: ["follow"]
      operationId: unfollowUser
      summary: Unfollow a user
      description: |
        This request is used to unfollow a user.
        The userId of the user to unfollow is passed as a path parameter.
        The userId of the user who is unfollowing is taken from the bearer token.
        The response will retun the new list of followers of the user.
      responses:
        "200":
          description: The new list of followers of the user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/userCollection'
        "404": #user not found
          $ref: '#/components/responses/NotFound'
        "500": #server error
          $ref: '#/components/responses/InternalServerError'
        "401": #unauthorized
          $ref: '#/components/responses/Unauthorized'
...
  /users/{userId}/banned/{bannedId}:
    description: A user banned by a user
    parameters:
        - $ref: '#/components/parameters/userId'
        - $ref: '#/components/parameters/bannedId'

    put:
      tags: ["ban"]
      operationId: banUser
      summary: Ban a user
      description: |
        This request is used to ban a user.
        The userId of the user to ban is passed as a path parameter.
        The userId of the user who is banning is taken from the bearer token.
        The response will retun the new list of banned users of the user.
      responses:
        "201":
          description: The new list of banned users of the user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/userCollection'
        "401": #unauthorized
          $ref: '#/components/responses/Unauthorized'
        "404": #user not found
          $ref: '#/components/responses/NotFound'
        "500": #server error
          $ref: '#/components/responses/InternalServerError'
    
    delete:
      tags: ["ban"]
      operationId: unbanUser
      summary: Unban a user
      description: |
        This request is used to unban a user.
        The userId of the user to unban is passed as a path parameter.
        The userId of the user who is unbanning is taken from the bearer token (its also in the path since one can only change is own ban list).
        The response will retun the new list of banned users of the user.
      responses:
        "200":
          description: The new list of banned users of the user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/userCollection'
        "401": #unauthorized
          $ref: '#/components/responses/Unauthorized'
        "500": #server error
          $ref: '#/components/responses/InternalServerError'
...
security:
  - bearerAuth: [] 